Add files via upload
In the project for our SQL course we had to create a balance sheet, cash flow statement and profit & loss statement. Even though that is not the real purpose of the language itself it was a nice and challenging activity.

I had the pleasure to work with my good friend Alvaro Flores Lorente on that!